migrate handler functions to a seperate file
